Control starting pathfinding & refactor
- control how many new paths per frame (with move all setting)
- random walk: generate targets such as the time to get there is controlled by config, so we can tune with the above

+++ src/ai/sparse_grid.rs
@@ -14,8 +14,8 @@ impl GridPos {
         Self { x, y }
     }
 
-    pub fn distance(&self, other: GridPos) -> u32 {
-        (((self.x - other.x).pow(2) + (self.y - other.y).pow(2)) as f32).sqrt() as u32
+    pub fn distance(&self, other: GridPos) -> f32 {
+        (((self.x - other.x).pow(2) + (self.y - other.y).pow(2)) as f32).sqrt()
     }
 
     pub const VN_OFFSETS: [GridPos; 4] = [
